requirements.txt

opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==2.0.0
Flask==3.0.3
aiohttp==3.9.5
aiortc==1.9.0
av==12.2.0
load-dotenv==0.1.0
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6

# opencv-python==4.5.3.56
# mediapipe==0.8.7.3
# numpy==1.21.2
flask[async]
